requirements.txt
Django==2.1.1
djangorestframework==3.8.2
psycopg2==2.7.5
gunicorn==19.3.0
